KAFKA-8558: Move KStreamImplJoin to its own class
